requirements.txt
# Core dependencies
google-genai>=1.16.1
streamlit>=1.45.1
python-dotenv>=1.1.0

# RAG and NLP (CPU optimized versions)
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers[torch]>=4.30.0  # Explicitly use torch backend

# Document processing
pypdf>=5.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
typing-extensions>=4.5.0

# Explicitly avoid TensorFlow
--no-deps tensorflow